extracted schema to its own file
